reconfigure application structure for testing
