+++ package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test": "jest",
     "start": "nodemon server.js",
     "debug": "nodemon --inspect server.js"
   },
@@ -24,6 +24,8 @@
   },
   "devDependencies": {
     "dotenv": "^16.1.4",
-    "nodemon": "^2.0.22"
+    "jest": "^29.6.1",
+    "nodemon": "^2.0.22",
+    "supertest": "^6.3.3"
   }
 }
